AutoPlaque-QSM-python/requirements.txt
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
nibabel>=5.1.0
Pillow>=10.0.0
scikit-image>=0.21.0
PyYAML>=6.0
